+++ source/math/libMath_plane.cpp
@@ -34,7 +34,7 @@ vec3 planeLineIntersection(vec3 _planePoint, vec3 _planeNormal, vec3 _linePoint,
         return vec3(0.0f, 0.0f, 0.0f);
     }
     
-    float64 t = (_planeNormal.dot(_planePoint) - _planeNormal.dot(_linePoint)) / pnDld;
+    float32 t = (_planeNormal.dot(_planePoint) - _planeNormal.dot(_linePoint)) / pnDld;
     
     vec3 ldSt = ld * t;
     vec3 lpPld = _linePoint + ldSt;
